lint

+++ src/storage/page/hash_table_block_page.cpp
@@ -47,7 +47,7 @@ template <typename KeyType, typename ValueType, typename KeyComparator>
 void HASH_TABLE_BLOCK_TYPE::Remove(slot_offset_t bucket_ind) {
   // only set readable to 0
   size_t char_idx, bit_idx;
-  unsigned short one = 0x01;
+  unsigned int one = 0x01;
   char_idx = bucket_ind / 8;
   bit_idx = bucket_ind % 8;
   // if occupied
@@ -59,7 +59,7 @@ void HASH_TABLE_BLOCK_TYPE::Remove(slot_offset_t bucket_ind) {
 template <typename KeyType, typename ValueType, typename KeyComparator>
 bool HASH_TABLE_BLOCK_TYPE::IsOccupied(slot_offset_t bucket_ind) const {
   size_t char_idx, bit_idx;
-  unsigned short one = 0x01;
+  unsigned int one = 0x01;
   char_idx = bucket_ind / 8;
   bit_idx = bucket_ind % 8;
   return (occupied_[char_idx] >> bit_idx) & one;
@@ -68,7 +68,7 @@ bool HASH_TABLE_BLOCK_TYPE::IsOccupied(slot_offset_t bucket_ind) const {
 template <typename KeyType, typename ValueType, typename KeyComparator>
 bool HASH_TABLE_BLOCK_TYPE::IsReadable(slot_offset_t bucket_ind)  const {
   size_t char_idx, bit_idx;
-  unsigned short one = 0x01;
+  unsigned int one = 0x01;
   char_idx = bucket_ind / 8;
   bit_idx = bucket_ind % 8;
   return (readable_[char_idx] >> bit_idx) & one;
